tools/citation_tools/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "citation-tools"
version = "0.1.0"
description = "Convert BibTeX entries to formatted citations using CSL styles"
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
authors = [
    {name = "Henrik Holm Justesen"}
]
keywords = ["bibtex", "citations", "csl", "pandoc", "bibliography"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Text Processing :: Markup",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "bibtexparser>=1.2.0",
    "PyYAML>=5.4",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
clipboard = [
    "pyperclip>=1.8.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
    "black>=22.0",
    "ruff>=0.0.200",
]

[project.scripts]
cite = "citation_tools.cli:main"
cite-setup = "citation_tools.setup_config:setup"

[tool.setuptools]
packages = ["citation_tools"]

[tool.setuptools.package-data]
citation_tools = ["styles/*.csl"]

# Post-install hook to run configuration setup
[tool.setuptools.cmdclass]

[tool.black]
line-length = 100
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
target-version = "py37"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
